+++ src/mergeSx.ts
@@ -1,6 +1,7 @@
 import type { SxProps } from "@mui/system";
 
-const ensureArray = (sx: SxProps<any>) => (Array.isArray(sx) ? sx : [sx]);
+type PureSx<T extends object> = Exclude<SxProps<T>, ReadonlyArray<any>>;
+type SxAsArray<T extends object> = Array<PureSx<T>>;
 
 /**
  * @desc Combines multiple SxProps
@@ -11,11 +12,22 @@ const ensureArray = (sx: SxProps<any>) => (Array.isArray(sx) ? sx : [sx]);
  * @requires SxProps from MUI 5.1.0 or higher
  * @see https://github.com/mui/material-ui/releases/tag/v5.1.0
  * @link https://github.com/mui/material-ui/pull/29297
+ * @since v0.1.4 using for..of instead of reducer for performance reasons
  */
 export const mergeSx = <T extends object>(
   ...styles: (SxProps<T> | false | undefined)[]
-): SxProps<T> =>
-  styles.reduce<SxProps<T>>(
-    (agg, sx) => ensureArray(agg).concat(ensureArray(sx || [])),
-    []
-  );
+): SxProps<T> => {
+  const capacitor: SxAsArray<T> = [];
+  for (const sx of styles) {
+    if (sx) {
+      if (Array.isArray(sx)) {
+        for (const sub of sx as SxAsArray<T>) {
+          capacitor.push(sub);
+        }
+      } else {
+        capacitor.push(sx as PureSx<T>);
+      }
+    }
+  }
+  return capacitor;
+};